requirements.txt
opencv-python
mediapipe==0.10.21
numpy
pillow
customtkinter